requirements.txt
fastapi
uvicorn
asyncpg
psycopg2-binary
celery
redis
